Refactor and reduce code duplication
 - Bug Fixes in UI
 - Toggle to exclude immutable contracts from funds at risk accounting
