+++ packages/protocolbeat/src/defidisco/DependencyInventoryBreakdown.tsx
@@ -1,271 +1,31 @@
-import { useMutation, useQueryClient } from '@tanstack/react-query'
-import { useEffect, useState } from 'react'
+import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query'
+import { useMemo, useState } from 'react'
 import { useParams } from 'react-router-dom'
-import { updateContractTag, updateFunction } from '../api/api'
+import { getProject, updateContractTag, updateFunction } from '../api/api'
 import type {
+  AdminModuleScore,
   DependencyModuleScore,
   Impact,
-  LetterGrade,
   Likelihood,
 } from '../api/types'
+import { buildProxyTypeMap } from '../apps/discovery/defidisco/proxyTypeUtils'
 import { usePanelStore } from '../apps/discovery/store/panel-store'
 import { useContractTags } from '../hooks/useContractTags'
+import {
+  computeWorstGrade,
+  getGradeBadgeStyles,
+  getGradeColor,
+  getLikelihoodColor,
+  ImpactPicker,
+  impactToScore,
+  isZeroAddress,
+  LikelihoodPicker,
+  OwnerSection,
+} from './scoringShared'
 
 interface DependencyInventoryBreakdownProps {
   score: DependencyModuleScore
-}
-
-/**
- * Get semantic color class for a letter grade
- */
-function getGradeColor(grade: LetterGrade): string {
-  switch (grade) {
-    case 'AAA':
-    case 'AA':
-    case 'A':
-      return 'text-green-400'
-    case 'BBB':
-    case 'BB':
-    case 'B':
-      return 'text-yellow-400'
-    case 'CCC':
-    case 'CC':
-    case 'C':
-      return 'text-orange-400'
-    case 'D':
-      return 'text-red-400'
-  }
-}
-
-/**
- * Get inline styles for grade badge
- */
-function getGradeBadgeStyles(grade: LetterGrade): {
-  backgroundColor: string
-  borderColor: string
-  color: string
-} {
-  switch (grade) {
-    case 'AAA':
-    case 'AA':
-    case 'A':
-      return {
-        backgroundColor: 'rgba(20, 83, 45, 0.5)', // green-900/50
-        borderColor: 'rgba(34, 197, 94, 0.3)', // green-500/30
-        color: '#4ade80', // green-400
-      }
-    case 'BBB':
-    case 'BB':
-    case 'B':
-      return {
-        backgroundColor: 'rgba(113, 63, 18, 0.5)', // yellow-900/50
-        borderColor: 'rgba(234, 179, 8, 0.3)', // yellow-500/30
-        color: '#facc15', // yellow-400
-      }
-    case 'CCC':
-    case 'CC':
-    case 'C':
-      return {
-        backgroundColor: 'rgba(124, 45, 18, 0.5)', // orange-900/50
-        borderColor: 'rgba(249, 115, 22, 0.3)', // orange-500/30
-        color: '#fb923c', // orange-400
-      }
-    case 'D':
-      return {
-        backgroundColor: 'rgba(127, 29, 29, 0.5)', // red-900/50
-        borderColor: 'rgba(239, 68, 68, 0.3)', // red-500/30
-        color: '#f87171', // red-400
-      }
-  }
-}
-
-/**
- * Get color value for impact level (inline style)
- */
-function getImpactColor(impact: string): string {
-  switch (impact) {
-    case 'critical':
-      return '#c084fc' // purple-400
-    case 'high':
-      return '#f87171' // red-400
-    case 'medium':
-      return '#fbbf24' // yellow-400
-    case 'low':
-      return '#10b981' // green-500
-    default:
-      return '#9ca3af' // gray-400
-  }
-}
-
-/**
- * Get color value for likelihood level (inline style)
- */
-function getLikelihoodColor(likelihood: string): string {
-  switch (likelihood) {
-    case 'high':
-      return '#f87171' // red-400
-    case 'medium':
-      return '#fb923c' // orange-400
-    case 'low':
-      return '#10b981' // green-500
-    case 'mitigated':
-      return '#60a5fa' // blue-400
-    default:
-      return '#9ca3af' // gray-400 (unassigned)
-  }
-}
-
-/**
- * Convert Impact to score string for API
- */
-function impactToScore(
-  impact: Impact,
-): 'low-risk' | 'medium-risk' | 'high-risk' | 'critical' {
-  switch (impact) {
-    case 'low':
-      return 'low-risk'
-    case 'medium':
-      return 'medium-risk'
-    case 'high':
-      return 'high-risk'
-    case 'critical':
-      return 'critical'
-  }
-}
-
-/**
- * Impact inline editor dropdown
- */
-function ImpactPicker({
-  currentImpact,
-  onUpdate,
-}: {
-  currentImpact: Impact
-  onUpdate: (impact: Impact) => void
-}) {
-  const [isOpen, setIsOpen] = useState(false)
-  const impactOptions: Impact[] = ['low', 'medium', 'high', 'critical']
-
-  const handleSelect = (impact: Impact) => {
-    onUpdate(impact)
-    setIsOpen(false)
-  }
-
-  return (
-    <div className="relative inline-block">
-      <button
-        onClick={(e) => {
-          e.stopPropagation()
-          setIsOpen(!isOpen)
-        }}
-        className="rounded border border-coffee-600 bg-coffee-700 px-2 py-0.5 text-xs capitalize hover:bg-coffee-600"
-        style={{ color: getImpactColor(currentImpact) }}
-      >
-        {currentImpact}
-      </button>
-
-      {isOpen && (
-        <>
-          <div
-            className="fixed inset-0 z-40"
-            onClick={() => setIsOpen(false)}
-          />
-          <div className="absolute right-0 z-50 mt-1 flex min-w-[120px] flex-col gap-2 rounded border border-coffee-600 bg-coffee-800 p-2 shadow-xl">
-            <div className="font-semibold text-coffee-300 text-xs">Impact</div>
-            {impactOptions.map((imp) => (
-              <button
-                key={imp}
-                className={`rounded border border-coffee-600 px-2 py-1 text-xs capitalize ${
-                  currentImpact === imp
-                    ? 'bg-coffee-600'
-                    : 'bg-coffee-700 hover:bg-coffee-600'
-                }`}
-                onClick={() => handleSelect(imp)}
-              >
-                {imp}
-              </button>
-            ))}
-          </div>
-        </>
-      )}
-    </div>
-  )
-}
-
-/**
- * Likelihood inline editor dropdown
- */
-function LikelihoodPicker({
-  currentLikelihood,
-  onUpdate,
-}: {
-  currentLikelihood: Likelihood
-  onUpdate: (likelihood: Likelihood) => void
-}) {
-  const [isOpen, setIsOpen] = useState(false)
-  const [selectedLikelihood, setSelectedLikelihood] =
-    useState<Likelihood>(currentLikelihood)
-  const likelihoodOptions: Likelihood[] = ['high', 'medium', 'low', 'mitigated']
-
-  // Sync internal state when prop changes
-  useEffect(() => {
-    setSelectedLikelihood(currentLikelihood)
-  }, [currentLikelihood])
-
-  const handleApply = () => {
-    onUpdate(selectedLikelihood)
-    setIsOpen(false)
-  }
-
-  const handleOpen = (e: React.MouseEvent) => {
-    e.stopPropagation()
-    setIsOpen(!isOpen)
-  }
-
-  return (
-    <div className="relative inline-block">
-      <button
-        onClick={handleOpen}
-        className="rounded border border-coffee-600 bg-coffee-700 px-2 py-0.5 text-xs capitalize hover:bg-coffee-600"
-        style={{ color: getLikelihoodColor(currentLikelihood) }}
-      >
-        {currentLikelihood}
-      </button>
-
-      {isOpen && (
-        <>
-          <div
-            className="fixed inset-0 z-40"
-            onClick={() => setIsOpen(false)}
-          />
-          <div className="absolute right-0 z-50 mt-1 flex min-w-[120px] flex-col gap-2 rounded border border-coffee-600 bg-coffee-800 p-2 shadow-xl">
-            <div className="font-semibold text-coffee-300 text-xs">
-              Likelihood
-            </div>
-            {likelihoodOptions.map((lik) => (
-              <button
-                key={lik}
-                className={`rounded border border-coffee-600 px-2 py-1 text-xs capitalize ${
-                  selectedLikelihood === lik
-                    ? 'bg-coffee-600'
-                    : 'bg-coffee-700 hover:bg-coffee-600'
-                }`}
-                onClick={() => setSelectedLikelihood(lik)}
-              >
-                {lik}
-              </button>
-            ))}
-            <button
-              className="w-full rounded border border-coffee-600 bg-coffee-700 px-2 py-1 text-xs hover:bg-coffee-600"
-              onClick={handleApply}
-            >
-              Apply
-            </button>
-          </div>
-        </>
-      )}
-    </div>
-  )
+  adminScore?: AdminModuleScore
 }
 
 /**
@@ -287,35 +47,14 @@ function DependencySection({
   const [isExpanded, setIsExpanded] = useState(false)
   const selectGlobal = usePanelStore((state) => state.select)
 
-  // Calculate worst grade among all functions for this dependency
-  const worstGrade =
-    dependency.functions.length > 0
-      ? dependency.functions.reduce((worst: LetterGrade, func: any) => {
-          const gradeValues: Record<LetterGrade, number> = {
-            AAA: 10,
-            AA: 9,
-            A: 8,
-            BBB: 7,
-            BB: 6,
-            B: 5,
-            CCC: 4,
-            CC: 3,
-            C: 2,
-            D: 1,
-          }
-          return gradeValues[func.grade] < gradeValues[worst]
-            ? func.grade
-            : worst
-        }, dependency.functions[0].grade)
-      : null
-
+  const worstGrade = computeWorstGrade(dependency.functions)
   const badgeStyles = worstGrade ? getGradeBadgeStyles(worstGrade) : null
 
   return (
-    <div className="mb-2 ml-4">
+    <div className="mb-1 ml-4">
       <button
         onClick={() => setIsExpanded(!isExpanded)}
-        className="flex w-full items-center gap-2 rounded p-2 text-left transition-colors hover:bg-coffee-800/30"
+        className="flex w-full items-center gap-2 rounded px-2 py-1 text-left transition-colors hover:bg-coffee-800/30"
       >
         <span className="text-coffee-400 text-xs">
           {isExpanded ? '▼' : '▶'}
@@ -348,6 +87,7 @@ function DependencySection({
           onUpdate={(likelihood) =>
             onUpdateLikelihood(dependency.dependencyAddress, likelihood)
           }
+          allowUnscored={true}
         />
         <span className="ml-2 text-coffee-400 text-xs">
           ({dependency.functions.length} function
@@ -358,24 +98,34 @@ function DependencySection({
       {isExpanded && (
         <ul className="mt-2 ml-8 space-y-1.5">
           {dependency.functions.map((func: any, idx: number) => {
-            const likelihoodColor = getLikelihoodColor(dependency.likelihood)
-            const gradeBadgeStyles = getGradeBadgeStyles(func.grade)
+            const likelihoodColor = getLikelihoodColor(
+              dependency.likelihood || '',
+            )
+            const gradeBadgeStyles = func.grade
+              ? getGradeBadgeStyles(func.grade)
+              : null
 
             return (
               <li
                 key={idx}
                 className="flex items-center gap-2 text-coffee-300 text-xs"
               >
-                <span
-                  className="inline-block rounded border px-1.5 py-0.5 font-mono text-xs"
-                  style={{
-                    backgroundColor: gradeBadgeStyles.backgroundColor,
-                    borderColor: gradeBadgeStyles.borderColor,
-                    color: gradeBadgeStyles.color,
-                  }}
-                >
-                  {func.grade}
-                </span>
+                {gradeBadgeStyles ? (
+                  <span
+                    className="inline-block rounded border px-1.5 py-0.5 font-mono text-xs"
+                    style={{
+                      backgroundColor: gradeBadgeStyles.backgroundColor,
+                      borderColor: gradeBadgeStyles.borderColor,
+                      color: gradeBadgeStyles.color,
+                    }}
+                  >
+                    {func.grade}
+                  </span>
+                ) : (
+                  <span className="inline-block px-1.5 py-0.5 text-coffee-500 text-xs">
+                    -
+                  </span>
+                )}
                 <button
                   onClick={() => selectGlobal(func.contractAddress)}
                   className="cursor-pointer font-medium text-coffee-200 transition-colors hover:text-blue-400"
@@ -397,7 +147,7 @@ function DependencySection({
                 />
                 <span className="text-coffee-500">, Likelihood: </span>
                 <span style={{ color: likelihoodColor }}>
-                  {dependency.likelihood}
+                  {dependency.likelihood || 'unscored'}
                 </span>
                 <span className="text-coffee-500">)</span>
               </li>
@@ -411,15 +161,29 @@ function DependencySection({
 
 /**
  * Dependency Inventory Breakdown Component
- * Displays breakdown of dependencies by external contract
+ * Displays breakdown of dependencies by external contract,
+ * including external owners from the admin breakdown.
  */
 export function DependencyInventoryBreakdown({
   score,
+  adminScore,
 }: DependencyInventoryBreakdownProps) {
   const { project } = useParams()
   const queryClient = useQueryClient()
   const { data: contractTags } = useContractTags(project!)
   const gradeColor = getGradeColor(score.grade)
+
+  // Fetch project data for proxy type information (needed for external owners)
+  const { data: projectData } = useQuery({
+    queryKey: ['projects', project],
+    queryFn: () => getProject(project!),
+    enabled: !!project,
+  })
+
+  const proxyTypeMap = useMemo(
+    () => buildProxyTypeMap(projectData),
+    [projectData],
+  )
 
   // Mutation for updating likelihood
   const updateLikelihoodMutation = useMutation({
@@ -432,8 +196,6 @@ export function DependencyInventoryBreakdown({
     }) => {
       if (!project) throw new Error('Project not found')
 
-      // Get existing tag to preserve other attributes
-      // IMPORTANT: Keep the eth: prefix for matching!
       const existingTag = contractTags?.tags.find(
         (tag) =>
           tag.contractAddress.toLowerCase() === contractAddress.toLowerCase(),
@@ -447,7 +209,6 @@ export function DependencyInventoryBreakdown({
       })
     },
     onSuccess: () => {
-      // Invalidate queries to refetch data
       queryClient.invalidateQueries({ queryKey: ['contract-tags', project] })
       queryClient.invalidateQueries({ queryKey: ['v2-score', project] })
     },
@@ -480,7 +241,6 @@ export function DependencyInventoryBreakdown({
       })
     },
     onSuccess: () => {
-      // Invalidate queries to refetch data
       queryClient.invalidateQueries({ queryKey: ['functions', project] })
       queryClient.invalidateQueries({ queryKey: ['v2-score', project] })
     },
@@ -494,41 +254,118 @@ export function DependencyInventoryBreakdown({
     updateImpactMutation.mutate({ contractAddress, functionName, impact })
   }
 
-  // Count functions across all dependencies
-  const totalFunctionCount = score.breakdown
-    ? score.breakdown.reduce((sum, dep) => sum + dep.functions.length, 0)
-    : 0
+  // Toggle for showing/hiding immutable dependencies (default: on)
+  const [showImmutable, setShowImmutable] = useState(true)
+
+  // Extract external owners from admin breakdown
+  const allExternalOwners = useMemo(() => {
+    if (!adminScore?.breakdown || !contractTags?.tags) return []
+
+    return adminScore.breakdown.filter((admin) => {
+      return contractTags.tags.some(
+        (tag) =>
+          tag.contractAddress.toLowerCase() ===
+            admin.adminAddress.toLowerCase() && tag.isExternal,
+      )
+    })
+  }, [adminScore, contractTags])
+
+  // Filter immutable/revoked external owners based on toggle
+  const isImmutableOrRevoked = (admin: any) =>
+    proxyTypeMap.get(admin.adminAddress.toLowerCase()) === 'immutable' ||
+    isZeroAddress(admin.adminAddress)
+
+  const displayedExternalOwners = useMemo(() => {
+    if (showImmutable) {
+      // Show all, but put immutable/revoked at the end
+      const mutable = allExternalOwners.filter((a) => !isImmutableOrRevoked(a))
+      const immutable = allExternalOwners.filter((a) => isImmutableOrRevoked(a))
+      return [...mutable, ...immutable]
+    }
+    return allExternalOwners.filter((a) => !isImmutableOrRevoked(a))
+  }, [allExternalOwners, proxyTypeMap, showImmutable])
+
+  const hasImmutableExternalOwners = useMemo(() => {
+    return allExternalOwners.some((a) => isImmutableOrRevoked(a))
+  }, [allExternalOwners, proxyTypeMap])
+
+  // Regular dependencies (from call graph / function dependencies)
+  const regularDeps = score.breakdown || []
+
+  // Count functions across regular dependencies
+  const depFunctionCount = regularDeps.reduce(
+    (sum, dep) => sum + dep.functions.length,
+    0,
+  )
+
+  // Count functions across displayed external owners
+  const extOwnerFunctionCount = displayedExternalOwners.reduce(
+    (sum, admin) => sum + admin.functions.length,
+    0,
+  )
+
+  const totalFunctionCount = depFunctionCount + extOwnerFunctionCount
+  const totalContractCount = regularDeps.length + displayedExternalOwners.length
+  const hasAnyEntries =
+    regularDeps.length > 0 || allExternalOwners.length > 0
 
   return (
     <div className="text-coffee-300">
-      {/* Main header - non-expandable, consistent with other inventory items */}
+      {/* Main header */}
       <div className="flex items-center justify-between">
         <span className="font-medium">Dependencies:</span>
-        <span>
-          {score.inventory}{' '}
-          <span className={`font-semibold ${gradeColor}`}>
-            (Grade: {score.grade})
+        <span className="flex items-center gap-2">
+          {hasImmutableExternalOwners && (
+            <label className="flex cursor-pointer items-center gap-1.5 text-xs text-coffee-400">
+              <input
+                type="checkbox"
+                checked={showImmutable}
+                onChange={(e) => setShowImmutable(e.target.checked)}
+                className="h-3 w-3 cursor-pointer accent-coffee-500"
+              />
+              Show immutable
+            </label>
+          )}
+          <span>
+            {score.inventory}{' '}
+            <span className={`font-semibold ${gradeColor}`}>
+              (Grade: {score.grade})
+            </span>
           </span>
         </span>
       </div>
 
-      {/* Dependency breakdown - always shown */}
+      {/* Dependency breakdown */}
       <div className="mt-3 ml-2">
-        {!score.breakdown || score.breakdown.length === 0 ? (
+        {!hasAnyEntries ? (
           <p className="ml-4 text-coffee-400 text-xs">
             No functions with dependencies on external contracts
           </p>
         ) : (
           <>
             <p className="mb-3 ml-4 text-coffee-400 text-xs">
-              {totalFunctionCount} function{totalFunctionCount !== 1 ? 's' : ''}{' '}
-              using {score.breakdown.length} external contract
-              {score.breakdown.length !== 1 ? 's' : ''}
+              {totalFunctionCount} function
+              {totalFunctionCount !== 1 ? 's' : ''} using{' '}
+              {totalContractCount} external contract
+              {totalContractCount !== 1 ? 's' : ''}
             </p>
-            {score.breakdown.map((dep) => (
+
+            {/* Regular dependencies */}
+            {regularDeps.map((dep) => (
               <DependencySection
                 key={dep.dependencyAddress}
                 dependency={dep}
+                onUpdateLikelihood={handleUpdateLikelihood}
+                onUpdateImpact={handleUpdateImpact}
+              />
+            ))}
+
+            {/* External owners (rendered with OwnerSection for full tags/funds) */}
+            {displayedExternalOwners.map((admin) => (
+              <OwnerSection
+                key={admin.adminAddress}
+                admin={admin}
+                proxyType={proxyTypeMap.get(admin.adminAddress.toLowerCase())}
                 onUpdateLikelihood={handleUpdateLikelihood}
                 onUpdateImpact={handleUpdateImpact}
               />
