+++ packages/protocolbeat/src/defidisco/V2ScoringSection.tsx
@@ -1,6 +1,6 @@
 import { useQuery } from '@tanstack/react-query'
 import { getV2Score } from '../api/api'
-import type { LetterGrade, ModuleScore } from '../api/types'
+import type { AdminModuleScore, LetterGrade, ModuleScore } from '../api/types'
 import { AdminsInventoryBreakdown } from './AdminsInventoryBreakdown'
 import { DependencyInventoryBreakdown } from './DependencyInventoryBreakdown'
 import { FunctionBreakdown } from './FunctionBreakdown'
@@ -119,6 +119,7 @@ export function V2ScoringSection({ project }: V2ScoringSectionProps) {
             <FunctionBreakdown score={scoreData.inventory.functions} />
             <DependencyInventoryBreakdown
               score={scoreData.inventory.dependencies}
+              adminScore={scoreData.inventory.admins}
             />
             <AdminsInventoryBreakdown score={scoreData.inventory.admins} />
           </div>
